# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product API",
        "description": "API для управления товарами, арендой и покупками в системе Laravel Product",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить список всех товаров",
                "description": "Возвращает список всех доступных товаров в системе. Требуется авторизация.",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Список товаров получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создать новый товар",
                "description": "Создает новый товар в системе. Требуется роль администратора.",
                "operationId": "storeProduct",
                "requestBody": {
                    "description": "Данные для создания товара",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товар создан успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Не удалось создать товар",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав доступа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить товар по ID",
                "description": "Возвращает данные конкретного товара по его идентификатору. Требуется авторизация.",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновить товар",
                "description": "Обновляет данные существующего товара. Требуется роль администратора.",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления товара",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар обновлен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Не удалось обновить товар",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав доступа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удалить товар",
                "description": "Удаляет товар из системы. Требуется роль администратора.",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар удален успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Не удалось удалить товар",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав доступа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Ответ с ошибкой",
                "description": "Стандартный ответ API при ошибке",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Товар не найден"
                    },
                    "errors": {
                        "description": "Детали ошибок (опционально)",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Ответ с ошибками валидации",
                "description": "Ответ API при ошибках валидации",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Ошибка валидации данных"
                    },
                    "errors": {
                        "description": "Детали ошибок валидации",
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Имя обязательно"
                                ]
                            },
                            "purchase_price": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Цена покупки должна быть числом"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "Успешный ответ",
                "description": "Стандартный успешный ответ API",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Операция выполнена успешно"
                    },
                    "data": {
                        "description": "Данные ответа (если необходимо)",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Товар",
                "description": "Модель товара",
                "required": [
                    "id",
                    "name",
                    "purchase_price",
                    "rent_price_per_hour",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор товара",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название товара",
                        "type": "string",
                        "example": "Игровая консоль PlayStation 5"
                    },
                    "description": {
                        "description": "Описание товара",
                        "type": "string",
                        "example": "Новейшая игровая консоль от Sony с 4K поддержкой",
                        "nullable": true
                    },
                    "purchase_price": {
                        "description": "Цена покупки товара",
                        "type": "integer",
                        "example": 50000
                    },
                    "rent_price_per_hour": {
                        "description": "Цена аренды товара за час",
                        "type": "integer",
                        "example": 500
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    }
                },
                "type": "object"
            },
            "ProductRequest": {
                "title": "Запрос для создания/обновления товара",
                "description": "Данные для создания или обновления товара",
                "properties": {
                    "name": {
                        "description": "Название товара",
                        "type": "string",
                        "example": "Игровая консоль PlayStation 5"
                    },
                    "description": {
                        "description": "Описание товара",
                        "type": "string",
                        "example": "Новейшая игровая консоль от Sony с 4K поддержкой",
                        "nullable": true
                    },
                    "purchase_price": {
                        "description": "Цена покупки товара",
                        "type": "number",
                        "format": "integer",
                        "minimum": 0,
                        "example": 50000
                    },
                    "rent_price_per_hour": {
                        "description": "Цена аренды товара за час",
                        "type": "number",
                        "format": "integer",
                        "minimum": 0,
                        "example": 500
                    }
                },
                "type": "object"
            },
            "StoreProductRequest": {
                "required": [
                    "name",
                    "purchase_price",
                    "rent_price_per_hour"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ProductRequest"
                    }
                ]
            },
            "UpdateProductRequest": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ProductRequest"
                    }
                ]
            },
            "ProductResponse": {
                "title": "Ответ с товаром",
                "description": "Стандартный ответ API с данными товара",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Товар получен успешно"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "ProductListResponse": {
                "title": "Ответ со списком товаров",
                "description": "Стандартный ответ API со списком товаров",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Товары получены успешно"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Аутентификация через Laravel Sanctum токены",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}